Fix DMR Tier III voice-grant LCN parsed from wrong bits (#639)
The TV/PV voice-grant CSBK parser used the wrong field layout: it read
the "LCN" from the last payload octet, which is actually the low byte of
the 24-bit source (subscriber) address. The decoded LCN therefore changed
with every transmitting radio, even on a system with only a couple of
fixed channels.

Per ETSI TS 102 361-4 (cross-checked against the dsd-neo reference
decoder), the channel-grant content leads with a 12-bit Logical Physical
Channel Number, followed by the target then source addresses. Parse the
LPCN from the leading payload bits and the addresses from their correct
octets, and widen the DMR LCN from 7 to 12 bits end-to-end (parser,
band-plan resolver, autoconfig learner, config, events, persistence).

This unbreaks both operator-supplied dmr_band_plan tables and the LCN
autoconfig learner, which previously never converged because it observed
a fresh LCN on every call. Encrypted/Emergency are no longer derived from
a grant (the ETSI grant carries no service-options octet; DMR privacy is
signaled in the voice Link Control).

Adds a regression guard asserting two grants for the same channel but
different source addresses decode to the same LCN.

+++ internal/events/bus.go
@@ -254,7 +254,7 @@ type DecodeError struct {
 type DMRGrantObserved struct {
 	System    string
 	ColorCode uint8
-	LCN       uint8
+	LCN       uint16
 	Timeslot  uint8
 	GroupID   uint32
 	SourceID  uint32
@@ -282,7 +282,7 @@ type DMRBandPlanLearned struct {
 // DMRBandPlanLCN is one explicit LCN→downlink-frequency entry in a
 // learned irregular (table) band plan.
 type DMRBandPlanLCN struct {
-	LCN    uint8
+	LCN    uint16
 	FreqHz uint32
 }
